adjust marked
