+++ src/components/KymaCompanion/components/Chat/messages/markedExtension.tsx
@@ -7,6 +7,41 @@ export const UI5Renderer = {
     return <CodePanel text={tokens} />;
   },
 
+  blockquote(tokens: string) {
+    return (
+      <blockquote
+        style={{
+          fontStyle: 'italic',
+          borderLeft: '1px dashed black',
+          margin: '1em 0',
+          paddingLeft: '1em',
+        }}
+      >
+        {tokens}
+      </blockquote>
+    );
+  },
+
+  text(tokens: string) {
+    return <Text className="text">{tokens}</Text>;
+  },
+
+  list(items: any[]) {
+    return (
+      <>
+        <ul
+          className="sap-margin-begin-tiny sap-padding-x-tiny"
+          style={{
+            lineHeight: 1.5,
+            listStyleType: 'disc',
+          }}
+        >
+          {items}
+        </ul>
+      </>
+    );
+  },
+
   codespan(tokens: string) {
     return <Text className="text highlighted">{tokens}</Text>;
   },
@@ -23,6 +58,7 @@ export const UI5Renderer = {
       </Title>
     );
   },
+
   link(href: string, title: string) {
     return (
       <Link href={href} target="_blank">
